+++ front/lib/api/workos/organization.ts
@@ -1,0 +1,67 @@
+import type { Organization as WorkOSOrganization } from "@workos-inc/node";
+import { GeneratePortalLinkIntent } from "@workos-inc/node";
+
+import { getWorkOS } from "@app/lib/api/workos/utils";
+import { WorkOSPortalIntent } from "@app/lib/types/workos";
+import logger from "@app/logger/logger";
+import type { Result, WorkspaceType } from "@app/types";
+import { Err, Ok } from "@app/types";
+
+export function createWorkOSOrganization({
+  workspace,
+}: {
+  workspace: WorkspaceType;
+}): Result<Promise<WorkOSOrganization>, Error> {
+  if (workspace.workOSOrganizationId) {
+    return new Err(
+      new Error("A WorkOS organization already exists for this workspace.")
+    );
+  }
+
+  try {
+    const organization = getWorkOS().organizations.createOrganization({
+      name: workspace.name,
+      metadata: { workspaceSId: workspace.sId },
+    });
+
+    return new Ok(organization);
+  } catch (error) {
+    logger.error(error, "Failed to create WorkOS organization");
+    return new Err(new Error("Failed to create WorkOS organization"));
+  }
+}
+
+// Mapping WorkOSPortalIntent to GeneratePortalLinkIntent,
+// as we can't use the WorkOSPortalIntent enum on any Client-Side code.
+const INTENT_MAP: Record<WorkOSPortalIntent, GeneratePortalLinkIntent> = {
+  [WorkOSPortalIntent.SSO]: GeneratePortalLinkIntent.SSO,
+  [WorkOSPortalIntent.DSync]: GeneratePortalLinkIntent.DSync,
+  [WorkOSPortalIntent.DomainVerification]:
+    GeneratePortalLinkIntent.DomainVerification,
+  [WorkOSPortalIntent.AuditLogs]: GeneratePortalLinkIntent.AuditLogs,
+  [WorkOSPortalIntent.LogStreams]: GeneratePortalLinkIntent.LogStreams,
+  [WorkOSPortalIntent.CertificateRenewal]:
+    GeneratePortalLinkIntent.CertificateRenewal,
+};
+
+export function generateWorkOSAdminPortalUrl({
+  organization,
+  workOSIntent,
+  returnUrl,
+}: {
+  organization: string;
+  workOSIntent: WorkOSPortalIntent;
+  returnUrl: string;
+}) {
+  const intent = INTENT_MAP[workOSIntent];
+
+  if (!intent) {
+    throw new Error(`Invalid intent: ${workOSIntent}`);
+  }
+
+  return getWorkOS().portal.generateLink({
+    organization,
+    intent,
+    returnUrl,
+  });
+}
