+++ front/lib/api/workos/user.ts
@@ -1,0 +1,72 @@
+import type {
+  AuthenticationResponse as WorkOSAuthenticationResponse,
+  User as WorkOSUser,
+} from "@workos-inc/node";
+import { unsealData } from "iron-session";
+import type { GetServerSidePropsContext, NextApiRequest } from "next";
+
+import config from "@app/lib/api/config";
+import type { RegionType } from "@app/lib/api/regions/config";
+import { getUserNicknameFromEmail, getWorkOS } from "@app/lib/api/workos/utils";
+import type { SessionWithUser } from "@app/lib/iam/provider";
+
+export type SessionCookie = {
+  sessionData: string;
+  organizationId?: string;
+  authenticationMethod: WorkOSAuthenticationResponse["authenticationMethod"];
+  region: RegionType;
+  workspaceId: string;
+};
+
+export async function getWorkOSSession(
+  req: NextApiRequest | GetServerSidePropsContext["req"]
+): Promise<SessionWithUser | undefined> {
+  const workOSSessionCookie = req.cookies["workos_session"];
+  if (workOSSessionCookie) {
+    const { sessionData, organizationId, authenticationMethod, workspaceId } =
+      await unsealData<SessionCookie>(workOSSessionCookie, {
+        password: config.getWorkOSCookiePassword(),
+      });
+
+    const session = getWorkOS().userManagement.loadSealedSession({
+      sessionData,
+      cookiePassword: config.getWorkOSCookiePassword(),
+    });
+
+    const r = await session.authenticate();
+
+    if (!r.authenticated) {
+      return undefined;
+    }
+
+    return {
+      type: "workos" as const,
+      sessionId: r.sessionId,
+      user: {
+        sid: r.user.id,
+        email: r.user.email,
+        email_verified: r.user.emailVerified,
+        name: r.user.email ?? "",
+        nickname: getUserNicknameFromEmail(r.user.email) ?? "",
+        sub: r.user.id,
+      },
+      // TODO(workos): Should we resolve the workspaceId and remove organizationId from here?
+      organizationId,
+      workspaceId,
+      isSSO: authenticationMethod === "SSO",
+      authenticationMethod,
+    };
+  }
+}
+
+// Store the region in the user's app_metadata to redirect to the right region.
+// A JWT Template includes this metadata in https://dust.tt/region (https://dashboard.workos.com/environment_01JGCT54YDGZAAD731M0GQKZGM/authentication/edit-jwt-template)
+export async function setRegionForUser(user: WorkOSUser, region: RegionType) {
+  // Update user metadata
+  await getWorkOS().userManagement.updateUser({
+    userId: user.id,
+    metadata: {
+      region,
+    },
+  });
+}
