+++ front/pages/api/w/[wId]/workos/sync.ts
@@ -1,7 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
-import { syncWorkOSDirectoriesForWorkspace } from "@app/lib/api/workos";
+import { syncWorkOSDirectoriesForWorkspace } from "@app/lib/api/workos/sync";
 import type { Authenticator } from "@app/lib/auth";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types";
